Upgrade lotus v0.5.4.
