+++ api/struct.go
@@ -18,7 +18,6 @@ import (
 	"github.com/filecoin-project/lotus/api"
 	"github.com/filecoin-project/lotus/build"
 	"github.com/filecoin-project/lotus/chain/types"
-	"github.com/filecoin-project/lotus/markets/utils"
 	"github.com/filecoin-project/lotus/node/modules/dtypes"
 	"github.com/filecoin-project/lotus/node/repo/importmgr"
 	"github.com/filecoin-project/sector-storage/ffiwrapper"
@@ -42,7 +41,6 @@ import (
 	"github.com/ipld/go-ipld-prime/traversal/selector"
 	"github.com/ipld/go-ipld-prime/traversal/selector/builder"
 	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/multiformats/go-multiaddr"
 	mh "github.com/multiformats/go-multihash"
 	"golang.org/x/xerrors"
 	"io"
@@ -133,7 +131,7 @@ func (a *API) ClientStartDeal(ctx context.Context, params *api.StartDealParams) 
 		return nil, xerrors.New("data doesn't fit in a sector")
 	}
 
-	providerInfo := utils.NewStorageProviderInfo(params.Miner, mi.Worker, mi.SectorSize, mi.PeerId, mi.Multiaddrs)
+	providerInfo := nodeapi.NewStorageProviderInfo(params.Miner, mi.Worker, mi.SectorSize, *mi.PeerId, mi.Multiaddrs)
 
 	log.Infof("Starting the deal")
 	dealStart := params.DealStartEpoch
@@ -257,7 +255,7 @@ func (a *API) ClientMinerQueryOffer(ctx context.Context, miner address.Address, 
 	}
 	rp := rm.RetrievalPeer{
 		Address: miner,
-		ID:      mi.PeerId,
+		ID:      *mi.PeerId,
 	}
 	return a.makeRetrievalQuery(ctx, rp, root, piece, rm.QueryParams{}), nil
 }
@@ -416,7 +414,7 @@ func (a *API) ClientRetrieve(ctx context.Context, order api.RetrievalOrder, ref 
 		}
 
 		order.MinerPeer = retrievalmarket.RetrievalPeer{
-			ID:      mi.PeerId,
+			ID:      *mi.PeerId,
 			Address: order.Miner,
 		}
 	}
@@ -520,30 +518,6 @@ func (a *API) ClientRetrieve(ctx context.Context, order api.RetrievalOrder, ref 
 	return files.WriteTo(file, ref.Path)
 }
 
-func NewStorageProviderInfo(address address.Address, miner address.Address, sectorSize abi.SectorSize, peer peer.ID, addrs []abi.Multiaddrs) storagemarket.StorageProviderInfo {
-	multiaddrs := make([]multiaddr.Multiaddr, 0, len(addrs))
-	if addrs == nil {
-		peerInfo, _ := nodeapi.NodeClient.UtilsAPI.NetFindPeer(context.Background(), peer)
-		multiaddrs = peerInfo.Addrs
-	} else {
-		for _, a := range addrs {
-			maddr, err := multiaddr.NewMultiaddrBytes(a)
-			if err != nil {
-				return storagemarket.StorageProviderInfo{}
-			}
-			multiaddrs = append(multiaddrs, maddr)
-		}
-	}
-
-	return storagemarket.StorageProviderInfo{
-		Address:    address,
-		Worker:     miner,
-		SectorSize: uint64(sectorSize),
-		PeerID:     peer,
-		Addrs:      multiaddrs,
-	}
-}
-
 func (a *API) ClientQueryAsk(ctx context.Context, p peer.ID, miner address.Address) (*storagemarket.SignedStorageAsk, error) {
 	log.Infof("Querying miner ask")
 	mi, err := a.StateMinerInfo(ctx, miner, types.EmptyTSK)
@@ -551,7 +525,7 @@ func (a *API) ClientQueryAsk(ctx context.Context, p peer.ID, miner address.Addre
 		return nil, xerrors.Errorf("failed getting miner info: %w", err)
 	}
 
-	info := NewStorageProviderInfo(miner, mi.Worker, mi.SectorSize, p, mi.Multiaddrs)
+	info := nodeapi.NewStorageProviderInfo(miner, mi.Worker, mi.SectorSize, p, mi.Multiaddrs)
 	signedAsk, err := a.SMDealClient.GetAsk(ctx, info)
 	if err != nil {
 		return nil, err
