+++ config/config.json
@@ -1,10 +1,10 @@
 {
   "Node" : {
     "Addr" : "/ip4/127.0.0.1/tcp/1234/ws",
-    "Token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJBbGxvdyI6WyJyZWFkIiwid3JpdGUiLCJzaWduIiwiYWRtaW4iXX0.v1n6zsl8--de000t9XZqCs-mFgYK9iMKNXljQHseql8"
+    "Token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJBbGxvdyI6WyJyZWFkIiwid3JpdGUiLCJzaWduIiwiYWRtaW4iXX0.i0fcD4TlFj9ZgS0BEHNdGm0RfCygYe-j4Zq_S-9Y8gI"
   },
   "Market" : {
     "Addr" : "/ip4/127.0.0.1/tcp/8888/ws",
-    "Token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJBbGxvdyI6WyJyZWFkIiwid3JpdGUiLCJzaWduIiwiYWRtaW4iXX0.v1n6zsl8--de000t9XZqCs-mFgYK9iMKNXljQHseql8"
+    "Token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJBbGxvdyI6WyJyZWFkIiwid3JpdGUiLCJzaWduIiwiYWRtaW4iXX0.i0fcD4TlFj9ZgS0BEHNdGm0RfCygYe-j4Zq_S-9Y8gI"
   }
 }
