+++ nodeapi/utils.go
@@ -5,6 +5,9 @@ package nodeapi
 
 import (
 	"context"
+	"github.com/filecoin-project/go-fil-markets/storagemarket"
+	"github.com/libp2p/go-libp2p-core/peer"
+	"github.com/multiformats/go-multiaddr"
 
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/lotus/api"
@@ -18,4 +21,28 @@ func GetStorageDeal(ctx context.Context, client Node, dealID abi.DealID, ts *typ
 
 func StateMinerInfo(ctx context.Context, client Node, ts *types.TipSet, maddr address.Address) (api.MinerInfo, error) {
 	return NodeClient.UtilsAPI.StateMinerInfo(ctx, maddr, ts.Key())
+}
+
+func NewStorageProviderInfo(address address.Address, miner address.Address, sectorSize abi.SectorSize, peer peer.ID, addrs []abi.Multiaddrs) storagemarket.StorageProviderInfo {
+	multiaddrs := make([]multiaddr.Multiaddr, 0, len(addrs))
+	if addrs == nil {
+		peerInfo, _ := NodeClient.UtilsAPI.NetFindPeer(context.Background(), peer)
+		multiaddrs = peerInfo.Addrs
+	} else {
+		for _, a := range addrs {
+			maddr, err := multiaddr.NewMultiaddrBytes(a)
+			if err != nil {
+				return storagemarket.StorageProviderInfo{}
+			}
+			multiaddrs = append(multiaddrs, maddr)
+		}
+	}
+
+	return storagemarket.StorageProviderInfo{
+		Address:    address,
+		Worker:     miner,
+		SectorSize: uint64(sectorSize),
+		PeerID:     peer,
+		Addrs:      multiaddrs,
+	}
 }
